+++ CHANGELOG.md
@@ -15,6 +15,7 @@ HumHub Changelog
 - Enh #8083: Store module version in DB
 - Enh #5141: Added CLI command to delete disabled or soft deleted users
 - Enh #8100: Move `static` and `themes` into `protected/humhub`
+- Enh #8129: Limit activity content length
 
 1.18.3 (TBD)
 ------------